requirements.txt
numpy==1.24.3
faiss-cpu==1.7.4
sentence-transformers==2.2.2
torch==2.0.1
datasets==2.13.0
groq==0.4.1
flask==2.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==2.3.6
requests==2.31.0
